+++ towhee/tests/engine/test_operator_context.py
@@ -15,37 +15,55 @@
 
 import unittest
 
-from towhee.engine.operator_context import OperatorContext
+from towhee.engine.operator_context import OperatorContext, OpStatus
 from towhee.dataframe import DataFrame, Variable
 from towhee.dag import OperatorRepr
+from towhee.engine.thread_pool_task_executor import ThreadPoolTaskExecutor
 
 from towhee.tests.test_util.dataframe_test_util import DfWriter
+from towhee.tests import CACHE_PATH
 
 
-class TestOperatorContext(unittest.TestCase):
-    """UT for `OperatorContext`
+class TestOperatorContextV2(unittest.TestCase):
+    """
+    Op ctx test
     """
 
-    def test_op_context(self):
+    def setUp(self):
+        self._task_exec = ThreadPoolTaskExecutor('op_ctx_test_executor',
+                                                 CACHE_PATH)
+        self._task_exec.start()
 
-        df_in = DataFrame('op_test_in')
-        df_out = DataFrame('op_test_out')
+    def tearDown(self):
+        if self._task_exec.is_alive():
+            self._task_exec.stop()
+            self._task_exec.join()
+
+    def _create_op_ctx(self):
+        df_in = DataFrame('op_test_in', {'num': {'type': 'int', 'index': 0}})
+        df_out = DataFrame('op_test_out', {'sum': {'type': 'int', 'index': 0}})
         dfs = {'op_test_in': df_in, 'op_test_out': df_out}
 
         op_repr = OperatorRepr(
-            name='mock_op',
-            function='mock_op',
-            init_args=None,
+            name='add_operator',
+            function='mock_operators/add_operator',
+            init_args={'factor': 5},
             inputs=[
-                {'name': 'k1', 'df': 'op_test_in', 'col': 0},
-                {'name': 'k2', 'df': 'op_test_in', 'col': 1}
+                {'name': 'num', 'df': 'op_test_in', 'col': 0},
             ],
             outputs=[{'df': 'op_test_out'}],
             iter_info={'type': 'map'}
         )
 
-        op = OperatorContext(op_repr, dfs)
-        self.assertEqual(len(op.pop_ready_tasks()), 0)
+        op_ctx = OperatorContext(op_repr, dfs)
+        self.assertEqual(op_ctx.name, 'add_operator')
+        self.assertEqual(op_ctx.status, OpStatus.NOT_RUNNING)
+        op_ctx.start(self._task_exec)
+        self.assertEqual(op_ctx.status, OpStatus.RUNNING)
+        return df_in, df_out, op_ctx
+
+    def test_op_ctx(self):
+        df_in, df_out, op_ctx = self._create_op_ctx()
 
         data = (Variable('int', 1), Variable(
             'str', 'test'), Variable('float', 0.1))
@@ -54,52 +72,38 @@ class TestOperatorContext(unittest.TestCase):
         t.set_sealed_when_stop()
         t.start()
         t.join()
+        op_ctx.join()
+        self.assertEqual(op_ctx.status, OpStatus.FINISHED)
+        df_out.seal()
+        map_iter = df_out.map_iter()
+        for item in map_iter:
+            self.assertEqual(item[0].value, 6)
 
-        count = 0
-        while True:
-            tasks = op.pop_ready_tasks(2)
-            if not op.has_tasks:
-                break
-            self.assertEqual(len(tasks), 2)
-            count += len(tasks)
-            self.assertEqual(op.num_ready_tasks, data_size - count)
+    def test_op_ctx_failed(self):
+        df_in, df_out, op_ctx = self._create_op_ctx()
 
-        self.assertEqual(count, data_size)
+        # Set errer data
+        data = (Variable('str', 'test'), )
+        data_size = 20
+        t = DfWriter(df_in, data_size, data=data)
+        t.set_sealed_when_stop()
+        t.start()
+        t.join()
+        op_ctx.join()
+        self.assertEqual(op_ctx.status, OpStatus.FAILED)
+        df_out.seal()
 
-    def test_op_context_multithread(self):
+        self.assertEqual(df_out.size, 0)
 
-        df_in = DataFrame('op_test_in')
-        df_out = DataFrame('op_test_out')
-        dfs = {'op_test_in': df_in, 'op_test_out': df_out}
-
-        op_repr = OperatorRepr(
-            name='mock_op',
-            function='mock_op',
-            init_args=None,
-            inputs=[
-                {'name': 'k1', 'df': 'op_test_in', 'col': 0},
-                {'name': 'k2', 'df': 'op_test_in', 'col': 1}
-            ],
-            outputs=[{'df': 'op_test_out'}],
-            iter_info={'type': 'map'}
-        )
-
-        op = OperatorContext(op_repr, dfs)
-        self.assertEqual(len(op.pop_ready_tasks()), 0)
+    def test_op_ctx_stop(self):
+        df_in, _, op_ctx = self._create_op_ctx()
 
         data = (Variable('int', 1), Variable(
             'str', 'test'), Variable('float', 0.1))
         data_size = 20
         t = DfWriter(df_in, data_size, data=data)
-        t.set_sealed_when_stop()
         t.start()
-
-        count = 0
-        while True:
-            tasks = op.pop_ready_tasks(5)
-            if not op.has_tasks:
-                break
-            self.assertLessEqual(len(tasks), 2)
-            count += len(tasks)
-
-        self.assertEqual(count, data_size)
+        t.join()
+        op_ctx.stop()
+        op_ctx.join()
+        self.assertEqual(op_ctx.status, OpStatus.FINISHED)
