+++ towhee/tests/engine/test_operator_pool.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 from towhee.operator import Operator
 from towhee.engine.operator_pool import OperatorPool
-from towhee.engine.task import Task
+from towhee.engine.operator_runner.runner_base import _OpInfo
 
 
 class TestOperatorPool(unittest.TestCase):
@@ -36,10 +36,12 @@ class TestOperatorPool(unittest.TestCase):
     def test_acquire_release(self):
 
         hub_op_id = 'mock_operators/add_operator'
-        task = Task('test', hub_op_id, {'factor': 0}, (1), 0)
 
-        # Acquire the operator.
-        op = self._op_pool.acquire_op(task)
+        op_info = _OpInfo('add_operator', hub_op_id, {'factor': 0})
+
+        op = self._op_pool.acquire_op(op_info.op_key,
+                                      op_info.hub_op_id,
+                                      op_info.op_args)
 
         # Perform some simple operations.
         self.assertTrue(isinstance(op, Operator))
@@ -47,8 +49,9 @@ class TestOperatorPool(unittest.TestCase):
 
         # Release and re-acquire the operator.
         self._op_pool.release_op(op)
-        op = self._op_pool.acquire_op(task)
-
+        op = self._op_pool.acquire_op(op_info.op_key,
+                                      op_info.hub_op_id,
+                                      op_info.op_args)
         # Perform more operations.
         self.assertEqual(op(-1).sum, -1)
         self.assertEqual(op(100).sum, 100)
